clean code: server controllers/testing/reset to use test_helper for testusers & testsenders, remove useless comments
