+++ server/test/login_api.test.js
@@ -3,10 +3,8 @@ const supertest = require('supertest')
 const helper = require('./test_helper')
 const app = require('../app')
 const api = supertest(app)
-const User = require('../models/user')
 
 beforeAll(async () => {
-  await User.deleteMany({})
   await supertest(app)
     .post('/api/testing/reset')
     .expect(201)
@@ -101,6 +99,5 @@ describe('When there is initially one admin-user and two user-users at db', () =
       .send(userdata)
       .expect(401)
       .expect('Content-Type', /application\/json/)
-
   })
 })
